+++ decidim-core/app/services/decidim/notification_generator_for_recipient.rb
@@ -11,13 +11,14 @@ module Decidim
     # event - A String with the name of the event.
     # event_class - The class that wraps the event, in order to decorate it.
     # resource - an instance of a class implementing the `Decidim::Resource` concern.
-    # recipient - the ID of the User that will receive the notification.
+    # recipient - the User that will receive the notification.
     # extra - a Hash with extra information to be included in the notification.
-    def initialize(event, event_class, resource, recipient_id, extra)
+    def initialize(event, event_class, resource, recipient, user_role, extra) # rubocop:disable Metrics/ParameterLists
       @event = event
       @event_class = event_class
       @resource = resource
-      @recipient_id = recipient_id
+      @recipient = recipient
+      @user_role = user_role
       @extra = extra
     end
 
@@ -42,14 +43,10 @@ module Decidim
         event_class: event_class,
         resource: resource,
         event_name: event,
-        extra: extra
+        extra: extra.merge(received_as: user_role)
       )
     end
 
-    attr_reader :event, :event_class, :resource, :recipient_id, :extra
-
-    def recipient
-      @recipient ||= User.find_by(id: recipient_id)
-    end
+    attr_reader :event, :event_class, :resource, :recipient, :user_role, :extra
   end
 end