+++ decidim-initiatives/app/commands/decidim/initiatives/create_initiative.rb
@@ -95,7 +95,7 @@ module Decidim
           event: "decidim.events.initiatives.initiative_created",
           event_class: Decidim::Initiatives::CreateInitiativeEvent,
           resource: initiative,
-          recipient_ids: initiative.author.followers.pluck(:id)
+          followers: initiative.author.followers
         )
       end