+++ decidim-initiatives/app/commands/decidim/initiatives/vote_initiative.rb
@@ -53,7 +53,7 @@ module Decidim
           event: "decidim.events.initiatives.initiative_endorsed",
           event_class: Decidim::Initiatives::EndorseInitiativeEvent,
           resource: @initiative,
-          recipient_ids: @initiative.author.followers.pluck(:id)
+          followers: @initiative.author.followers
         )
       end
 
@@ -68,7 +68,8 @@ module Decidim
           event: "decidim.events.initiatives.milestone_completed",
           event_class: Decidim::Initiatives::MilestoneCompletedEvent,
           resource: @initiative,
-          recipient_ids: @initiative.followers.pluck(:id),
+          affected_users: [@initiative.author],
+          followers: @initiative.followers - [@initiative.author],
           extra: {
             percentage: percentage
           }